separated handler from mainwindow
